requirements.txt
TTS
python-dotenv
